backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["patrashu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.0"
uvicorn = "0.27.1"
torch = {version = "2.0.1+cu117", source = "pytorch"}
torchvision = {version = "0.15.2+cu117", source = "pytorch"}
timm = "0.9.16"
transformers = "4.38.1"
efficientnet_pytorch = "0.7.1"
Pillow = "10.2.0"
numpy = "1.26.4"
pandas = "2.0.0"
opencv-python = "4.9.0.80"
python-multipart = "^0.0.9"
scikit-learn = "^1.4.1.post1"
ttach = "^0.0.3"
matplotlib = "^3.8.3"
pydicom = "^2.4.4"
pydantic-settings = "^2.2.1"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu117"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
